+++ src/Wheel.h
@@ -13,7 +13,7 @@ class Wheel : public WheelBase {
 
     Wheel();																	//default constructor.
 
-    virtual void move(float velocity);                                //velocity demanded radians per second.
+    virtual void move(double velocity);                                //velocity demanded radians per second.
     virtual void stop();															                //reset duty to 0 and direction to forward
     virtual void update();
